refactor(sqs): no longer tightly coupled to SQS client
Refactors SQS to decouple from the SQS client, instead providing
building blocks on *sqs.Message to start spans.
